feat: DetailModal moved to src/components, refactor modal styles, improve long-text editing
