+++ src/app/globals.css
@@ -39,3 +39,48 @@
   border: 1px solid #eee;
   border-radius: 8px;
 }
+
+/* Modal styles */
+.modalOverlay {
+  position: fixed; inset: 0;
+  background: rgba(0, 0, 0, 0.4);
+  display: flex; align-items: center; justify-content: center;
+  z-index: 1000;
+}
+
+.modalBox {
+  background: #ffffff;
+  width: min(1000px, 96vw);
+  max-height: 88vh;
+  overflow: auto;
+  border-radius: 12px;
+  padding: 24px;
+  box-shadow: 0 10px 30px rgba(38, 56, 220, 0.2);
+}
+
+.modalHeader {
+  display: flex; justify-content: space-between; align-items: center; margin-bottom: 16px;
+}
+.modalHeader h2 { margin: 0; font-size: 20px; }
+
+.modalGrid {
+  display: grid; grid-template-columns: 200px 1fr; gap: 12px;
+}
+.modalLabel { font-weight: 600; font-size: 14px; padding-top: 8px; }
+
+.modalInput {
+  width: 100%; min-height: 120px; resize: vertical;
+  padding: 10px; border: 1px solid #ddd; border-radius: 8px;
+  font-family: ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, "Liberation Mono", "Courier New", monospace;
+  font-size: 14px; line-height: 1.5;
+}
+
+.modalInputCompact {
+  width: 100%; height: 40px;
+  padding: 10px; border: 1px solid #ddd; border-radius: 8px;
+  font-size: 14px; line-height: 1.3;
+}
+
+.modalActions {
+  margin-top: 24px; display: flex; gap: 12px; justify-content: flex-end;
+}
